requirements.txt
pycryptoplus